\name{empiricalGameTheory-package}
\alias{empiricalGameTheory-package}
\alias{empiricalGameTheory}
\docType{package}
\title{
\packageTitle{empiricalGameTheory}
}
\description{
\packageDescription{empiricalGameTheory}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{empiricalGameTheory}
\packageIndices{empiricalGameTheory}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{empiricalGameTheory}

Maintainer: \packageMaintainer{empiricalGameTheory}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
